Add FoundationStereo TensorRT backend
