+++ tool/foundation_stereo_trt.py
@@ -1,6 +1,5 @@
 import argparse
 import asyncio
-import ctypes
 import logging
 import math
 import platform
@@ -8,9 +7,8 @@ import sys
 
 import cv2
 import numpy as np
-import tensorrt as trt
 from codetiming import Timer
-from cuda import cudart
+from tinynav.core.models_trt import FoundationStereoTRT
 
 try:
     import rclpy
@@ -26,180 +24,16 @@ except ImportError:
     ROS_AVAILABLE = False
     Node = object
 
-numpy_to_ctypes = {
-    np.dtype(np.float32): ctypes.c_float,
-    np.dtype(np.float16): ctypes.c_uint16,
-    np.dtype(np.int8): ctypes.c_int8,
-    np.dtype(np.uint8): ctypes.c_uint8,
-    np.dtype(np.int32): ctypes.c_int32,
-    np.dtype(np.int64): ctypes.c_int64,
-    np.dtype(np.bool_): ctypes.c_bool,
-}
-
-
-def disparity_to_depth(disparity: np.ndarray, baseline: float, focal_length: float) -> np.ndarray:
-    disparity = np.asarray(disparity, dtype=np.float32)
-    baseline = float(np.asarray(baseline).reshape(-1)[0])
-    focal_length = float(np.asarray(focal_length).reshape(-1)[0])
-
-    if baseline <= 0.0:
-        raise ValueError(f"baseline must be positive, got {baseline}")
-    if focal_length <= 0.0:
-        raise ValueError(f"focal_length must be positive, got {focal_length}")
-
-    depth = np.zeros_like(disparity, dtype=np.float32)
-    valid = np.isfinite(disparity) & (disparity > 0.0)
-    depth[valid] = (baseline * focal_length) / disparity[valid]
-    return depth
-
-
-class TRTBase:
-    def __init__(self, engine_path: str):
-        trt_logger = trt.Logger(trt.Logger.WARNING)
-        with open(engine_path, "rb") as f, trt.Runtime(trt_logger) as runtime:
-            self.engine = runtime.deserialize_cuda_engine(f.read())
-        self.context = self.engine.create_execution_context()
-        self.inputs, self.outputs, self.bindings, self.stream = self.allocate_buffers()
-
-    def allocate_buffers(self):
-        inputs = []
-        outputs = []
-        bindings = []
-        _, stream = cudart.cudaStreamCreate()
-
-        for i in range(self.engine.num_io_tensors):
-            name = self.engine.get_tensor_name(i)
-            shape = tuple(self.engine.get_tensor_shape(name))
-            dtype = np.dtype(trt.nptype(self.engine.get_tensor_dtype(name)))
-            ctype_dtype = numpy_to_ctypes[dtype]
-            is_input = self.engine.get_tensor_mode(name) == trt.TensorIOMode.INPUT
-
-            size = trt.volume(shape)
-            nbytes = size * dtype.itemsize
-
-            if "aarch64" in platform.machine():
-                ptr = cudart.cudaHostAlloc(nbytes, cudart.cudaHostAllocMapped)[1]
-                host_mem = np.ctypeslib.as_array((ctype_dtype * size).from_address(ptr))
-                host_mem = host_mem.view(dtype).reshape(shape)
-                device_ptr = cudart.cudaHostGetDevicePointer(ptr, 0)[1]
-            else:
-                ptr = cudart.cudaMallocHost(nbytes)[1]
-                host_mem = np.ctypeslib.as_array((ctype_dtype * size).from_address(ptr))
-                host_mem = host_mem.view(dtype).reshape(shape)
-                device_ptr = cudart.cudaMalloc(nbytes)[1]
-
-            bindings.append(int(device_ptr))
-            io = {
-                "name": name,
-                "host": host_mem,
-                "device": device_ptr,
-                "shape": shape,
-                "nbytes": nbytes,
-            }
-            if is_input:
-                inputs.append(io)
-            else:
-                outputs.append(io)
-
-        return inputs, outputs, bindings, stream
-
-    async def run_graph(self):
-        if "aarch64" not in platform.machine():
-            for inp in self.inputs:
-                cudart.cudaMemcpyAsync(
-                    inp["device"],
-                    inp["host"].ctypes.data,
-                    inp["nbytes"],
-                    cudart.cudaMemcpyKind.cudaMemcpyHostToDevice,
-                    self.stream,
-                )
-
-        for i in range(self.engine.num_io_tensors):
-            self.context.set_tensor_address(self.engine.get_tensor_name(i), self.bindings[i])
-        self.context.execute_async_v3(stream_handle=self.stream)
-
-        if "aarch64" not in platform.machine():
-            for out in self.outputs:
-                cudart.cudaMemcpyAsync(
-                    out["host"].ctypes.data,
-                    out["device"],
-                    out["nbytes"],
-                    cudart.cudaMemcpyKind.cudaMemcpyDeviceToHost,
-                    self.stream,
-                )
-
-        _, event = cudart.cudaEventCreate()
-        cudart.cudaEventRecord(event, self.stream)
-        while cudart.cudaEventQuery(event)[0] == cudart.cudaError_t.cudaErrorNotReady:
-            await asyncio.sleep(0)
-
-        return {out["name"]: out["host"].copy() for out in self.outputs}
-
-
-class FoundationStereoTRT(TRTBase):
-    def __init__(self, engine_path=f"/tinynav/tinynav/models/foundation_stereo_{platform.machine()}.plan"):
-        super().__init__(engine_path)
-        if len(self.inputs) != 2:
-            raise RuntimeError(f"Expected 2 inputs (left/right), got {len(self.inputs)}")
-        if len(self.outputs) != 1:
-            raise RuntimeError(f"Expected 1 output (disp), got {len(self.outputs)}")
-
-        self.left_idx = 0 if self.inputs[0]["name"] == "left" else 1
-        self.right_idx = 1 - self.left_idx
-        self.output_name = self.outputs[0]["name"]
-
-        self.net_h = int(self.inputs[self.left_idx]["shape"][2])
-        self.net_w = int(self.inputs[self.left_idx]["shape"][3])
-
-    def _to_three_channel_float(self, image: np.ndarray) -> np.ndarray:
-        if image.ndim == 2:
-            image = np.repeat(image[:, :, None], 3, axis=2)
-        elif image.ndim == 3 and image.shape[2] == 1:
-            image = np.repeat(image, 3, axis=2)
-        elif image.ndim == 3 and image.shape[2] >= 3:
-            image = image[:, :, :3]
-        else:
-            raise ValueError(f"Unsupported image shape: {image.shape}")
-        image = image.astype(np.float32, copy=False)
-        return np.transpose(image, (2, 0, 1))[None, ...]
-
-    async def infer(self, left_img, right_img, baseline, focal_length):
-        if left_img.shape[:2] != right_img.shape[:2]:
-            raise ValueError(f"Left/right shape mismatch: {left_img.shape} vs {right_img.shape}")
-
-        h_in, w_in = left_img.shape[:2]
-        if (h_in, w_in) != (self.net_h, self.net_w):
-            raise ValueError(
-                f"Input image size must match engine size ({self.net_h}, {self.net_w}), "
-                f"got ({h_in}, {w_in}). Rebuild engine for this resolution."
-            )
-
-        left_tensor = self._to_three_channel_float(left_img)
-        right_tensor = self._to_three_channel_float(right_img)
-
-        np.copyto(self.inputs[self.left_idx]["host"], left_tensor)
-        np.copyto(self.inputs[self.right_idx]["host"], right_tensor)
-
-        results = await self.run_graph()
-
-        disp_net = np.asarray(results[self.output_name], dtype=np.float32).reshape(self.net_h, self.net_w)
-        disp_net = np.clip(disp_net, 0.0, None)
-        disp = disp_net
-
-        depth = disparity_to_depth(disp, baseline, focal_length)
-        return disp, depth
-
-
 class FoundationStereoDepthNode(Node):
     def __init__(
         self,
-        engine_path=f"/tinynav/tinynav/models/foundation_stereo_{platform.machine()}.plan",
+        engine_path=f"/tinynav/tinynav/models/foundation_stereo_11-33-40_256x320_4_{platform.machine()}.plan",
         left_topic="/camera/camera/infra1/image_rect_raw",
         right_topic="/camera/camera/infra2/image_rect_raw",
         camera_info_topic="/camera/camera/infra2/camera_info",
         keyframe_depth_topic="/keyframe/depth",
         pointcloud_topic="/keyframe/pointcloud",
-        frame_id="camera_foundation_stereo",
+        frame_id="camera",
         world_frame_id="world",
         publish_pointcloud=False,
         min_depth=0.05,
@@ -363,7 +197,7 @@ class FoundationStereoDepthNode(Node):
             depth_msg.header.stamp = left_msg.header.stamp
             depth_msg.header.frame_id = self.frame_id
             self.keyframe_depth_pub.publish(depth_msg)
-            self._publish_world_to_foundation_tf(left_msg.header.stamp)
+            #self._publish_world_to_foundation_tf(left_msg.header.stamp)
 
         if self.publish_pointcloud:
             with Timer(
@@ -383,14 +217,14 @@ def main(args=None):
     parser.add_argument(
         "--engine_path",
         type=str,
-        default=f"/tinynav/tinynav/models/foundation_stereo_{platform.machine()}.plan",
+        default=f"/tinynav/tinynav/models/foundation_stereo_11-33-40_256x320_4_{platform.machine()}.plan",
     )
     parser.add_argument("--left_topic", type=str, default="/camera/camera/infra1/image_rect_raw")
     parser.add_argument("--right_topic", type=str, default="/camera/camera/infra2/image_rect_raw")
     parser.add_argument("--camera_info_topic", type=str, default="/camera/camera/infra2/camera_info")
     parser.add_argument("--keyframe_depth_topic", type=str, default="/keyframe/depth")
     parser.add_argument("--pointcloud_topic", type=str, default="/keyframe/pointcloud")
-    parser.add_argument("--frame_id", type=str, default="camera_foundation_stereo")
+    parser.add_argument("--frame_id", type=str, default="camera")
     parser.add_argument("--world_frame_id", type=str, default="world")
     parser.add_argument("--publish_pointcloud", action="store_true")
     parser.add_argument("--min_depth", type=float, default=0.05)
